Completely restructured the design of Project 5
Added a LinkedList class to use nodes to help with data management.
Added a SongList class to store data for all the songs. The Glyph class
was broken up into the SongGUI and StatsGUI classes for simplification.
A separate LegendGUI class was used to simplify the DisplayWindow class.
Comparators were added to the Hobby, Major, and RegionPercent classes.
All classes removed were due to design simplification or became
redundant.